+++ include/ZBE/archetypes/implementations/SimpleWideMobileAPO.h
@@ -19,13 +19,13 @@ template <unsigned s>
 class SimpleWideMobileAPO : public SimpleWideMobile<s>, public ActivePhysicalObject {
   public:
     SimpleWideMobileAPO(uint64_t width, uint64_t actuatorsList, uint64_t collisionablesList) : SimpleWideMobile<s>(width), ActivePhysicalObject(actuatorsList, collisionablesList) {}
-    SimpleWideMobileAPO(std::initializer_list<double> position, uint64_t width, uint64_t actuatorsList, uint64_t collisionablesList) : SimpleWideMobile<s>(position, width), ActivePhysicalObject(actuatorsList, collisionablesList) {}
+    SimpleWideMobileAPO(std::initializer_list<int64_t> position, uint64_t width, uint64_t actuatorsList, uint64_t collisionablesList) : SimpleWideMobile<s>(position, width), ActivePhysicalObject(actuatorsList, collisionablesList) {}
     SimpleWideMobileAPO(Point<s> position, uint64_t width, uint64_t actuatorsList, uint64_t collisionablesList) : SimpleWideMobile<s>(position, width), ActivePhysicalObject(actuatorsList, collisionablesList) {}
     SimpleWideMobileAPO(Vector<s> velocity, uint64_t width, uint64_t actuatorsList, uint64_t collisionablesList) : SimpleWideMobile<s>(velocity ,width), ActivePhysicalObject(actuatorsList, collisionablesList) {}
-    SimpleWideMobileAPO(std::initializer_list<double> position, std::initializer_list<double> velocity, uint64_t width, uint64_t actuatorsList, uint64_t collisionablesList) : SimpleWideMobile<s>(position, velocity, width), ActivePhysicalObject(actuatorsList, collisionablesList) {}
-    SimpleWideMobileAPO(Point<s> position, std::initializer_list<double> velocity, uint64_t width, uint64_t actuatorsList, uint64_t collisionablesList) : SimpleWideMobile<s>(position, velocity, width), ActivePhysicalObject(actuatorsList, collisionablesList) {}
+    SimpleWideMobileAPO(std::initializer_list<int64_t> position, std::initializer_list<int64_t> velocity, uint64_t width, uint64_t actuatorsList, uint64_t collisionablesList) : SimpleWideMobile<s>(position, velocity, width), ActivePhysicalObject(actuatorsList, collisionablesList) {}
+    SimpleWideMobileAPO(Point<s> position, std::initializer_list<int64_t> velocity, uint64_t width, uint64_t actuatorsList, uint64_t collisionablesList) : SimpleWideMobile<s>(position, velocity, width), ActivePhysicalObject(actuatorsList, collisionablesList) {}
     SimpleWideMobileAPO(Point<s> position, Vector<s> velocity, uint64_t width, uint64_t actuatorsList, uint64_t collisionablesList) : SimpleWideMobile<s>(position, velocity, width), ActivePhysicalObject(actuatorsList, collisionablesList) {}
-    SimpleWideMobileAPO(std::initializer_list<double> position, Vector<s> velocity, uint64_t width, uint64_t actuatorsList, uint64_t collisionablesList) : SimpleWideMobile<s>(position, velocity ,width), ActivePhysicalObject(actuatorsList, collisionablesList) {}
+    SimpleWideMobileAPO(std::initializer_list<int64_t> position, Vector<s> velocity, uint64_t width, uint64_t actuatorsList, uint64_t collisionablesList) : SimpleWideMobile<s>(position, velocity ,width), ActivePhysicalObject(actuatorsList, collisionablesList) {}
 
     virtual ~SimpleWideMobileAPO() {}
 };
